Fix running=false to not start the docker image on docker restart
When performing a systemctl restart docker, all images even the ones
with running=false are being started. This is unexpected behaviour.
The systemctl restart is needed when using docker in combination with
firewalld. After every firewalld restart, docker needs a restart in
order to make sure the containers keep working. The problem is fixed by
disabling the service in systemd. This makes sure the container isn't
respawned at a systemctl docker restart.